Introduce configs; enable fractional movement
Add centralized configuration (aiConfig.ts, gameConfig.ts, config/index.ts) and refactor movement/AI to support fractional, time-based movement. UNIT_STATS now derives moveSpeed from GAME_CONFIG.SERVER_TICK_RATE and includes movement docs. GameRoom uses GAME_CONFIG for simulation tick interval and spawn offsets/defaults. GameObjectSchema and UnitSchema position/target fields changed to float32 to allow sub-tile positions. AIBehaviorSystem now imports AI_CONFIG, uses AI_CONFIG values (wanderReplanInterval, maxWanderDistance), adds a small tolerance for target checks, and switches wandering to accumulate moveProgress (moving a tile when progress >= 1.0) for smooth movement across ticks.

+++ server/src/schema/GameObjectSchema.ts
@@ -14,10 +14,10 @@ export class GameObjectSchema extends Schema {
   @type("uint8")
   type: GameObjectType = GameObjectType.Castle;
 
-  @type("uint16")
+  @type("float32")
   x: number = 0;
 
-  @type("uint16")
+  @type("float32")
   y: number = 0;
 
   @type("string")
